nueva tabla en db para mascotas revision bug al crear mascota con peso en decimal

+++ app/src/main/java/com/dadm/appblackdog/models/UiPetForm.kt
@@ -38,7 +38,7 @@ data class UiPetForm(
             "name" to name,
             "breedId" to breedId,
             "ageRangeId" to ageRangeId,
-            "weight" to weight.toLong(),
+            "weight" to weight.toDouble(),
             "measureUnitId" to measureUnitId,
             "birthdate" to birthdateTimeStamp,
             "description" to description,
